+++ app/Factory/CompanyGatewayFactory.php
@@ -21,6 +21,8 @@ class CompanyGatewayFactory
         $company_gateway = new CompanyGateway;
         $company_gateway->company_id = $company_id;
         $company_gateway->user_id = $user_id;
+        $company_gateway->require_billing_address = false;
+        $company_gateway->require_shipping_address = false;
         // $company_gateway->fees_and_limits = new FeesAndLimits;
         
         return $company_gateway;