+++ app/Factory/InvoiceFactory.php
@@ -38,10 +38,10 @@ class InvoiceFactory
         $invoice->tax_rate2 = 0;
         $invoice->tax_name3 = '';
         $invoice->tax_rate3 = 0;
-        $invoice->custom_value1 = 0;
-        $invoice->custom_value2 = 0;
-        $invoice->custom_value3 = 0;
-        $invoice->custom_value4 = 0;
+        $invoice->custom_value1 = '';
+        $invoice->custom_value2 = '';
+        $invoice->custom_value3 = '';
+        $invoice->custom_value4 = '';
         $invoice->amount = 0;
         $invoice->balance = 0;
         $invoice->paid_to_date = 0;