+++ app/Mail/DownloadBackup.php
@@ -1,4 +1,13 @@
 <?php
+/**
+ * Invoice Ninja (https://invoiceninja.com).
+ *
+ * @link https://github.com/invoiceninja/invoiceninja source repository
+ *
+ * @copyright Copyright (c) 2021. Invoice Ninja LLC (https://invoiceninja.com)
+ *
+ * @license https://opensource.org/licenses/AAL
+ */
 
 namespace App\Mail;
 
@@ -9,7 +18,6 @@ use Illuminate\Queue\SerializesModels;
 
 class DownloadBackup extends Mailable
 {
-    // use Queueable, SerializesModels;
 
     public $file_path;
 
@@ -30,14 +38,15 @@ class DownloadBackup extends Mailable
         $company = Company::where('company_key', $this->company->company_key)->first();
 
         return $this->from(config('mail.from.address'), config('mail.from.name'))
-                    ->subject(ctrans('texts.download_backup_subject'))
+                    ->subject(ctrans('texts.download_backup_subject', ['company' => $company->present()->name()]))
                     ->markdown(
                         'email.admin.download_files',
                         [
                             'url' => $this->file_path,
                             'logo' => $company->present()->logo,
                             'whitelabel' => $company->account->isPaid() ? true : false,
-                            'settings' => $company->settings
+                            'settings' => $company->settings,
+                            'greeting' => $company->present()->name(),
                         ]
                     );
     }