+++ app/Models/BaseModel.php
@@ -13,12 +13,14 @@ namespace App\Models;
 
 use App\DataMapper\ClientSettings;
 use App\DataMapper\CompanySettings;
+use App\Jobs\Entity\CreateEntityPdf;
 use App\Utils\Traits\MakesHash;
 use App\Utils\Traits\UserSessionAttributes;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\ModelNotFoundException as ModelNotFoundException;
 use Illuminate\Support\Carbon;
+use Illuminate\Support\Facades\Storage;
 
 
 /**
@@ -199,4 +201,5 @@ class BaseModel extends Model
 
         return $formatted_number;
     }
+
 }