+++ app/Models/Company.php
@@ -443,12 +443,12 @@ class Company extends BaseModel
 
     public function domain()
     {
-        if (Ninja::isNinja()) {
+        if (Ninja::isHosted()) {
 
             if($this->portal_mode == 'domain')
                 return $this->portal_domain;
 
-            return "https://{$this->subdomain}" . config('ninja.app_domain');
+            return "https://{$this->subdomain}." . config('ninja.app_domain');
         }
 
         return config('ninja.app_url');