+++ resources/lang/en/texts.php
@@ -4250,6 +4250,7 @@ $LANG = array(
     'new_login_description' => 'You recently logged in to your Invoice Ninja account from a new location or device:<br><br><b>IP:</b> :ip<br><b>Time:</b> :time<br><b>Email:</b> :email',
     'download_backup_subject' => 'Your company backup is ready for download',
     'contact_details' => 'Contact Details',
+    'download_backup_subject' => ':company backup is ready for download',
 );
 
 return $LANG;