+++ resources/views/email/admin/download_files.blade.php
@@ -7,15 +7,20 @@
 
 @endslot
 
-@slot('greeting')
-@endslot
 
+@if(isset($greeting))
+<p style="padding-top:20px">{{ $greeting }}</p>
+@endif
+
+<p style="padding-top:20px">
 @lang('texts.download_timeframe')
+</p>
 
-
+<p style="padding-top:20px">
     @component('email.components.button', ['url' => $url])
         @lang('texts.download')
     @endcomponent
+</p>
 
 @slot('signature')
     InvoiceNinja (contact@invoiceninja.com)