+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased (daily channel)](https://github.com/invoiceninja/invoiceninja/tree/v5-develop)
 - Add Cache-control: no-cache to prevent overaggressive caching of assets
 - Improved labelling in the settings (client portal)
+- Client portal: Multiple accounts access improvements (#5703)
 
 ## [v5.1.56-release](https://github.com/invoiceninja/invoiceninja/releases/tag/v5.1.56-release)
 ## Fixed: